+++ src/index.js
@@ -1,1 +1,0 @@
-export { toINR, toINRWords } from './currencyFormatter'